% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\docType{data}
\name{Microbiome_dat}
\alias{Microbiome_dat}
\title{Taxonomy Reads and Associated Phenotypes}
\format{A list with 4 objects:
\describe{
  \item{pheno}{A data frame including two phenotype variables from 80 observations}
  \item{counts}{A matrix containing 40 taxa's read counts}
  \item{batch}{A numeric vector indicates the batch labels of the
  corresponding samples}
  \item{continuous}{A vector indicating whether the input phenotypes variables
  are continuous or not. 0 stands for discrete and 1 represents the
  continous.}
}}
\usage{
dat
}
\description{
A dataset containing the read counts of 40 taxa and their associated
phenotypes. The variables are as follows:
}
\keyword{datasets}
