% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDMMA.R
\name{VBatch}
\alias{VBatch}
\title{Visualize batch effect with principal coordinate analysis}
\usage{
VBatch(Y, batch, main_variable = NULL, method = "bray")
}
\arguments{
\item{Y}{Y A data.frame for taxonomic read counts with the rows representing clades
and columns representing taxa.}

\item{batch}{A numeric vector labeling the batch for each sample.}

\item{main_variable}{Optional. A vector containing the main effect variable. Only for
categorical main effect variable.The function will generate a figure for each catagory.}

\item{method}{A string indicating which method should be used to calculate the distance matrix for
principal coordinate analysis.}
}
\value{
The function returns a list containing plot objects of principal coordinate analysis figures.
}
\description{
Visualize batch effect with principal coordinate analysis
}
\examples{
require(SummarizedExperiment)
data(Microbiome_dat)
col_data=colData(Microbiome_dat)
pheno <- data.frame(col_data$V1, col_data$V2)
counts <- t(assay(Microbiome_dat))
batch <- col_data[,3]
figure <- VBatch(counts, batch = batch, main_variable = pheno[,1], method = "bray")
print(figure[[1]])
print(figure[[2]])
}
